// generated-contracts/models/blockchain-operation-list-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DeFi API (API users)
 * API surface available via API key authentication
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BlockchainOperationResponseDto } from './blockchain-operation-response-dto';
import {
    BlockchainOperationResponseDtoFromJSON,
    BlockchainOperationResponseDtoFromJSONTyped,
    BlockchainOperationResponseDtoToJSON,
    BlockchainOperationResponseDtoToJSONTyped,
} from './blockchain-operation-response-dto';

/**
 * 
 * @export
 * @interface BlockchainOperationListResponseDto
 */
export interface BlockchainOperationListResponseDto {
    /**
     * Total number of items across all pages
     * @type {number}
     * @memberof BlockchainOperationListResponseDto
     */
    total: number;
    /**
     * Current page number (starts from 1)
     * @type {number}
     * @memberof BlockchainOperationListResponseDto
     */
    page: number;
    /**
     * Number of items per page
     * @type {number}
     * @memberof BlockchainOperationListResponseDto
     */
    pageSize: number;
    /**
     * 
     * @type {Array<BlockchainOperationResponseDto>}
     * @memberof BlockchainOperationListResponseDto
     */
    items: Array<BlockchainOperationResponseDto>;
}

/**
 * Check if a given object implements the BlockchainOperationListResponseDto interface.
 */
export function instanceOfBlockchainOperationListResponseDto(value: object): value is BlockchainOperationListResponseDto {
    if (!('total' in value) || value['total'] === undefined) return false;
    if (!('page' in value) || value['page'] === undefined) return false;
    if (!('pageSize' in value) || value['pageSize'] === undefined) return false;
    if (!('items' in value) || value['items'] === undefined) return false;
    return true;
}

export function BlockchainOperationListResponseDtoFromJSON(json: any): BlockchainOperationListResponseDto {
    return BlockchainOperationListResponseDtoFromJSONTyped(json, false);
}

export function BlockchainOperationListResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BlockchainOperationListResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'total': json['total'],
        'page': json['page'],
        'pageSize': json['pageSize'],
        'items': ((json['items'] as Array<any>).map(BlockchainOperationResponseDtoFromJSON)),
    };
}

export function BlockchainOperationListResponseDtoToJSON(json: any): BlockchainOperationListResponseDto {
    return BlockchainOperationListResponseDtoToJSONTyped(json, false);
}

export function BlockchainOperationListResponseDtoToJSONTyped(value?: BlockchainOperationListResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'total': value['total'],
        'page': value['page'],
        'pageSize': value['pageSize'],
        'items': ((value['items'] as Array<any>).map(BlockchainOperationResponseDtoToJSON)),
    };
}